+++ nodalia-cards.manifest.js
@@ -1,10 +1,10 @@
 export default {
-  "pkgVersion": "1.1.1-alpha.2",
-  "contentSha256_12": "86a32e832945",
+  "pkgVersion": "1.1.1-alpha.3",
+  "contentSha256_12": "6fcdc4a96d04",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
-  "hacsFile": "nodalia-cards-1.1.1-alpha.2.js"
+  "hacsFile": "nodalia-cards-1.1.1-alpha.3.js"
 };
-export const pkgVersion = "1.1.1-alpha.2";
-export const contentSha256_12 = "86a32e832945";
+export const pkgVersion = "1.1.1-alpha.3";
+export const contentSha256_12 = "6fcdc4a96d04";
 export const file = "nodalia-cards.bundle.js";